+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - garuda: dual monitor configuration for hyprland
 - dev: `dependencies-update` & `dependencies-outdated` make tasks
 - debian: preferred packages install and purge script
+- debian: 1 day system log rotation
 
 ### Changed
 - git: restructure multiple configs, using relative include file paths